readyset-client: Extend update_controller_state
Extend the update_controller_state fn of AuthorityControl to allow us to
store the schema ReplicationOffset separately from ControllerState.

This will allow us to recover the schema replication offset if there is
a backwards incompatible change to the serialization format of
ControllerState.

This is implemebnted by having callers of update_controller_state
provide a function that allows the AuthorityControl to extract the
replication offset to store separately. This is a bit strange, but these
opaque functions at this level are (as far as I can tell) due to there
otherwise being either a cyclic dependency between readyset-client and
readyset-server, or else having server-specific types exposed too high
in the dependency graph.

This commit also implements the atomic storing of the schema replication
offset for our different Authority variants.

Change-Id: Ib0e117d4ad8e1588238617bedf3ffa8274335ae3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6374
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>