server: Restore schema replication offset with caches
If we fail to deserialize the ControllerState due to a backwards
incompatible change in ControllerState serialization format, we don't
want to have to resnapshot the tables. The tables themselves store the
replication offset associated with them, but there is also a schema
replication offset that is used to track changes that are unrelated to
any particular table "e.g. create custom type".

This change re-loads this schema offset from a place we stashed it off
to the side separately from the ControllerState.

It also contains a change to how we wait for lsns since the test
introduced for this behavior uncovered an edge case where we could be
stuck "Catching up" if we don't see any activity on the replication slot
after a restart.

Change-Id: I030b046c3e180a4a23632579066ffb1a5f7246e2