server: Restore schema replication offset with caches
If we fail to deserialize the ControllerState due to a backwards
incompatible change in ControllerState serialization format, we don't
want to have to resnapshot the tables. The tables themselves store the
replication offset associated with them, but there is also a schema
replication offset that is used to track changes that are unrelated to
any particular table "e.g. create custom type".

This change re-loads this schema offset from a place we stashed it off
to the side separately from the ControllerState.

Refs: REA-2931 #216
Change-Id: I46da45142b8c5bf716217a5f8fbe1d15462cabc9