adapter: Remove arc_unwrap_or_clone unstable feature
Rust unstable feature `arc_unwrap_or_clone` is only used in the adapter,
and acts as a minor optimization. Fortunately, there exists a way to
acheieve the same behavior without needing unstable APIs, if a bit
more verbose. Once the feature stabilizes, we should feel free to
switch back to the simpler syntax.

Change-Id: Ic69132956fcba7a953ef3155f6a6da503366846c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6376
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>