build: Expand sccache usage to `readyset-build` image
Layers of indirection had caused me to misjudge the scope of my
earlier sccache installation.  Specifically, our tests and many other
things run in the `readyset-build` image which does not have the ubuntu
builder base as an ancestor, and thus did not have a functioning
sccache.  This update deploys sccache more widely in the pipeline.

Change-Id: I7a2af6ad2d56bc76d916fea6c3c4550436488148
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6381
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>