readyset-psql: Mark types tests as #[slow]
The types tests are proptests that run in serial and make real database
calls. They are taking around 400 seconds to run, and are generally
unlikely to fail for most changes.

So running them in the nightly build only (when RUN_SLOW_TESTS=true is
set) will speed up normal builds significantly.

Change-Id: I7f5e1721a4f9c436ab7a66c50511840e6ada71b8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6384
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>