tests: Mark many serial tests as #[slow]
A lot of our tests are #[serial]ized because they contend on a real,
upstream database, including a replication slot for that database.

Since these tests aren't the long pole in the nightly pipeline, moving
them to there instead of running them with every build will drastically
speed up the common main build.

If we find that we are paying more cost by not finding failing tests
early enough, we can revert some of these, but my hypothesis is that it
will not be an issue.

Change-Id: I9bc4790e20f77c72620e725ae506f6ec9d681bdf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6386
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>