replicators: Wait for domains to be ready
At the beginning of replicator startup, we call replication_offsets(),
which itself calls into each domain to gather up the replication offsets
for the system. Before this commit, we retried this multiple times, but
with a noisy error log and a restart of the entire replicator loop.

Retrying just this command to wait for domains to be ready cleans up the
logs.

Refs: REA-2882
Change-Id: I415f541680ffe5123de3d54d2503719c643fde94
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6388
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>