replicators: Add no-op replication message after snapshotting
If there aren't any replication events while performing a re-snapshot,
we can get stuck in the 'catching up' phase because the resnapshot
slot's lsn is higher than the main slot, and there aren't any
replication events to advance the current lsn while catching up.

This commit adds a no-op replication event so that we can catch up even
in this case.

Change-Id: I911bcfb29bc0f3b2fc54a0696990e5791b2d5a45