adapter: Extract class for showing readyset status
This commit extracts the logic for obtaining the information for the
reuslt of a `show readyset status` call to a struct
ReadySetStatusReporter, and separates the formatting logic to a new
ReadySetStatus struct. This better encapsulates this logic and will
allow for easier re-use for a separate http/json formatting request.

Change-Id: I27caba35eda3ed8ee4a37a309cd5780957b10ffd