readyset-mysql: Extend 'show readyset status' test
This extends the test that asserts on the output format of 'show
readyset status' to fully capture the current format, and to make it so
that any update to the format (including additional lines), will cause a
test failure. We do this so that we can be sure an upcoming refactor of
this logic doesn't change the output format, nor will any future changes
without notifying us.

Change-Id: If91b84b6fdc8c8dcf9d820eb6826db43b1b7e287
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6397
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>