pipeline: Restrict framework test prep to "Framework Tests" group
The pre-command hook was trying to download the framework test generated
artifacts on every job.  That unnecessarily prolongs jobs (sometimes
over ten minutes!) that run after the framework tests but do not require
the artifacts.  This change will only perform the artifact download for
jobs in the "Framework Tests" group.

This is something of a brute force hammer.  A more precise solution
would be to have each generated framework test job download only the
artifact that it needs, and remove the download step from the
pre-command hook altogether.  However, do not let the Good become the
enemy of the Less Bad.

Change-Id: I6b0afad58a0e20268e0ca21ffc6cf0b40cc20913