quickstart: Add readyset_demo.sh script
This adds a bash script with the intention of automating the existing
demo. It provides some automation around running the docker containers
and seeding the sample database and an interactive prompt to guide a
user through the quickstart flow.

Along with the script is a testing script that uses `expect` to assert
on the main branches through the script, and pipelines to run the tests
and release the artifact when appropriate.

We run the demo script tests even if the demo script hasn't changed in
nightly to ensure that the demo hasn't broken because of a change in
readyset itself.

This commit also adds a Docker-in-Docker pipeline step to run the test,
and a Dockerfile to run that step locally for debugging. It was easier
to copy the simple install command into the step itself rather than have
a separate build step and to use the dockerfile in the pipeline itself.

Change-Id: I837a5b0eedcc50094a0f7513b4aa6f7ecf299984

Release-Note-Core: Added a convenient interactive script to run a
  ReadySet demo with. Check it out with
  `bash -c "$(curl -sSL https://launch.readyset.io/readyset_demo.sh)"`
Change-Id: Idb580722f269287c89d1cf4ad4204f83f95049b7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6445
Reviewed-by: Ron Hough <ron@readyset.io>
Tested-by: Buildkite CI