quickstart: Add readyset_demo.sh script
This adds a bash script with the intention of automating the existing
demo. It provides some automation around running the docker containers
and seeding the sample database and an interactive prompt to guide a
user through the quickstart flow.

This commit just adds the script itself to source control, and doesn't
include tests or deployment mechanisms yet.

Change-Id: I837a5b0eedcc50094a0f7513b4aa6f7ecf299984