adapter: Support `DROP CACHE <query string>`
This commit adds support for passing a query string to the `DROP CACHE`
SQL extension. The query string passed need not necessarily be the
rewritten, parameterized query string; similar to `CREATE CACHE FROM`, a
query with bound parameters (i.e. a query with literals instead of the
placeholders) will be parsed and rewritten before it is used to look up
an existing cache.

Release-Note-Core: Added support for passing a query string directly to
  `DROP CACHE`
Change-Id: Idcf04d8d1da3abc0d11cdcce1c1e24264053284a