buildkite: Improve handling of spot instance reclamation
The EC2 spot instances we use in CI are subject to sudden reclamation by
AWS.  The retry logic that was being used would auto-retry the job based
on exit status, but spot instance reclamation doesn't return a
consistent exit status.  The approach in this update will also retry if
a kill signal was sent to the job because the agent was stopped. That
should provide more accurate detection of spot instance reclamation.

See: https://buildkite.com/changelog/\
193-signal-and-signal-reason-in-automatic-retry-rules

Change-Id: I84f4e5c56bc21cff0d237913ff2e3e1cab22e192
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6462
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>