Demo: drop cache at the beginning for idempotency
If a user runs the demo script multiple times, the cache may already
exist, which will make the demonstration, which produces timings for
cache misses and cache hits, invalid.

This adds a 'drop cache' to the beginning of the interactive psql flow,
which makes the behavior consistent.

An extra select was added because we actually do 1 proxy, then 1
upquery after a drop/create cache cycle before seeing cache hits again.

Change-Id: I1f74a544d1130b9e12925addd2e8b3fa7eb2a764
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6469
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>