pipeline: Parallel psql generated logictests
This shards the psql generated logictests into 6 buckets (1 for roughly
each hour of curren test execution time).

The time savings nice, but this also allows a particular sub-test
to be retried more granularly, and also perhaps opens up the possibility
to use spot instances for these tests to save cost.

Note that this naive approach doesn't save anything on the time to pull
the repo and re-build the tests for each one. There may be something
smarter we can do there.

Change-Id: I5549de05a789af84ce60fe45e3fcadf66d960143