pipeline: Parallel psql generated logictests
This shards the psql generated logictests into 6 buckets (1 for roughly
each hour of current test execution time).

The time savings nice, but this also allows a particular sub-test to be
retried more granularly, and also uses spot instances for these tests to
save cost.

Note that this naive approach doesn't save anything on the time to pull
the repo and re-build the tests for each one. There may be something
smarter we can do there.

Change-Id: I95152b120ada8a5b79ab9343c8879b6ec867428b