demo: Rearrange testing logic to fix flake
Due to how the demo tests were structured and using regexes in a couple
of places, it was possible for one of the expected messages to be
globbed by the previous regex, which caused the next expect to time out.

This removes the problematic regex and also restructures the tests a bit
and puts the logic in a more linear fasion.

Change-Id: I7173060a666043676edfb49a43fe8e51d2cce209
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6478
Tested-by: Buildkite CI
Reviewed-by: Ron Hough <ron@readyset.io>