demo: Extend one-liner to allow postgres upstream
Now the one-liner can do either a demo or take an upstream connection as
user input and set up readyset pointing to that as well.

The bring-your-own branch of the script prompts for a connection string,
tests it (using docker psql, annoyingly, due to docker networking
peculiarities), then downloads and modifies the bring-your-own
docker-compose file with that connection string and starts it up. Then
it waits for snapshotting to finish. If that fails, we print the last
ERROR log we saw and link to troubleshooting pages. Otherwise we connect
to readyset and provide a list of readyset commands to get the user
started on their caching journey.

This commit updates the existing `expect` test to test the happy path of
this branch.

Change-Id: I01fea88813f9608952451064e355a71b1cf1abef