demo: add width to 'pv' command
This minor tweak adds a fixed width to the 'pv' output so that it
doesn't stretch across a very wide terminal.

Since we are suppressing debug output in the tests now, this also
installs pv on our docker container as it will no longer be overly noisy
in the test logs.

Change-Id: I4afdeb8ceedf88346c76f91b4cda795e9311e65c