adapter: Make `SHOW CACHES` persistent
Previously, the `SHOW CACHES` command pulled information from the query
status cache. This meant that after a restart, an existing cache would
only be visible via `SHOW CACHES` after the adapter handled a query
against that cache. `EXPLAIN CACHES` was the only source of truth for
the current set of caches, since it relied on querying the controller
to pull information from the expression registry.

This commit changes `SHOW CACHES` to use the expression registry, just
like `EXPLAIN CACHES`. It also adds the query ID to the expression
registry, since that data is needed in `SHOW CACHES`.

Release-Note-Core: Made the list of caches in `SHOW CACHES` persist
  after a restart
Change-Id: Ia49801e3eba4b806d949325b0fa71709dd35c2dc