tests: Add more retries to recreate database
The recreate database helper function drops a database and then attempts
to recreate it. We have seen some tests fail with a message hinting that
the previous database state (specifically an index into the system table
that holds the current database names) hasn't fully been removed:

> duplicate key value violates unique constraint
> \"pg_database_datname_index\"", detail: Some("Key (datname)=(noria)
> already exists.")

I'm not sure if this is a race condition or what, but I think adding
some extra sleeps between the DROP and CREATE and a retry loop to the
entire process is reasonable in an attempt to avoid test flakiness.

Change-Id: Iea4734c44a84587a82c2cc089147cdf4b29c2d9d