demo: Add grafana links to one-liner
This adds a link to our grafana dashboard to the one-liner.

It adds it in the guided tour section when we have just observed the
query speedup, and then again when we are connecting to readyset for
free-form exploration, and finally at the very end along with
reconnection instructions.

Because our quickstart redirect points to /current/ rather than /main/,
this also modifies the demo script to pull from github directly.

This also switches --wait-timeout for --wait as we found older versions
of docker don't have wait-timeout yet.

Change-Id: I7228dfa3d31e7aa7657de3585c9d1ab362678133
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6537
Reviewed-by: Ethan Donowitz <ethan@readyset.io>
Tested-by: Buildkite CI