adapter: Make `SHOW CACHES` persistent
Previously, `SHOW CACHES` used the query status cache as its source of
truth. This meant that after a restart, `SHOW CACHES` would not reflect
any caches that existed on the server until after the adapter received a
query against that cache.

This commit has `SHOW CACHES` use the expression registry as its source
of truth, which is what `EXPLAIN CACHES` uses. This enables the adapter
to display an accurate, up-to-date list of caches via `SHOW CACHES`,
even after a restart.

Refs: REA-3516, REA-3349
Release-Note-Core: Added the ability for the `SHOW CACHES` SQL extension
  to display an up-to-date list of caches, even after a restart
Change-Id: I822dd2cc1f061eb4ae22c9e0bc13534eb9c142be
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6546
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>