adapter+server: Refactor query IDs
Previously, the concept of a "query ID" was overloaded: the term was
used to describe the IDs generated by adapter-rewritten queries in the
query status cache *and* the IDs for expressions that were stored in the
expression registry. Both of these IDs had similar string
representations, making them look equivalent. We had two `QueryID` type
aliases (one for `String` and one for `u128`) and a separate `QueryId`
type, which was inconsistently used.

This commit refactors these type aliases and type into two types:
- `QueryId` (defined in `readyset-client`), which represents the ID of
  the adapter-rewritten query; and
- `ExprId` (defined in `readyset-server`), which represents the ID of an
  expression that is stored in the expression registry

This commit also makes usage of the `QueryId` type more consistent,
replacing parts of the code where we were just using `Strings` to
represent query IDs.

Change-Id: Icb405ea9a87ded8cfaf7840bc77f60c217430ae4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6545
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI