demo: auto swap docker localhost variants
The connection string that a user may provide (if running postgres in
docker locally) may not be one that our docker readyset can reach due to
docker networking quirks. If we fail to connect to the original
connection string and it has a localhost variant, try the mac and linux
docker 'localhost' variants that may work before asking for the user to
provide a different connection string.

Change-Id: I64799a9ace56c0a5e977387cbbe25baa198dd8d5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6563
Tested-by: Buildkite CI
Reviewed-by: Ron Hough <ron@readyset.io>