adapter: Fix views sync shutdown signal
Previously, we were not correctly waiting for the shutdown signal in the
`ViewsSynchronizer` -- when the `Interval::tick` future resolved, we
invoked `ViewsSynchronizer::poll` in a context where a shutdown signal
would be ignored. This commit ensures that the `tokio::select!`
invocation is invoked on a future that contains both the `tick` *and*
the `poll`, so a shutdown signal that occurs during the polling step
would still be respected.

Change-Id: Iad0aa05bab88db739679e05ca63fcccd12a5d475
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6565
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>