nom-sql: Reorder statement parsing
Trival reordering of the parsing of statmement in the primary
`parse()` function. This patch pushes the hot path statements
to the front of the parsing list (like SELECT, INSERT, and so on),
as we're more likely to hit those.

The actual performance benefit is in the noise-level as the parser
can bail fast when it finds a mismatch, but we save a few cycles
on the hot path for a trivial code reordering.

Fixes: REA-3923
Change-Id: I4e7c5de34179e7d2ee66c89da68a4777e07c9e60
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6586
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>
Reviewed-by: Vlad Rodionov <vlad.r@readyset.io>