nom-sql: Optimize SELECT statement parsing
In nom-sql, we parse for a compound SELECT statement before
parsing for a regular SELECT statement. Regular SELECT statements
are vastly more common in practice, and amounts to the same work:
compound SELECT parses for a SELECT, and then looks for a compound
operator (UNION, EXCEPT, or INTERSECT). Lacking that, it throws away
the parsed single SELECT, and then returns to the parser, just parse
the single SELECT all over again - duplicating work on the hot path.

This patch essentailly merges the compound SELECT and regular SELECT
parsing paths by not failing the compound parsing if a compound operator
is not found. In that case, it simply returns a `SelectStatement`.
Note: we must still parse for compound SELECTs first as parsing for a
single SELECT statement would leave the rest of the compound statement
unparsed (and thus we mis-identify a query).

Fixes: REA-3923
Release-Note-Core: Optimize SELECT statement parsing by
  eliminating deuplicated (and thrown away) work.
Change-Id: I00cfb54f74884c28b56b2c96655655ef8ec841c4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6587
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>