adapter: Make SHOW CACHES display query only
Previously, `SHOW CACHES` was displaying the entire `CREATE CACHE`
statement for each query as it existed in the expression registry. This
was not necessary: `SHOW CACHES` already includes the cache's name and
fallback behavior in the results. This commit makes a change so that
`SHOW CACHES` only displays the SELECT statement text.

Change-Id: Ic727f4c93800f8f29a3f0e0be1a1a039aa6ab9cc