demo: Allow passwordless connections
The demo is meant as a non-production way for folks to try out readyset,
and our (sane for production) default behavior is to not allow
passwordless connections, which we don't have control over for the demo.

This commit uses the `ALLOW_UNAUTHENTICATED_CONNECTIONS` flag to make
the default for the demo more permissive in case users are testing us
out with their dev postgres that doesn't have a password.

Change-Id: I57ec3aee8eae335954a12f5dc5ceec62c7ef967f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6619
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>