adapter: Add client connection open/close counters
We have a gauge for reporting the count of current client connections,
but we cannot know if those are the same connections or if the
client is thrashing connections (and the value just happens to
be the same). Hence, add two simple counters to record each
connection open and each close. That way we can see the rate of
opens and closes.

Change-Id: I0768c58a3703986dd0d69f77100ce51d2a54dc0c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6635
Reviewed-by: Vlad Rodionov <vlad.r@readyset.io>
Tested-by: Buildkite CI