nom-sql: Optional flamegraphs from benchmarks
Add functionality to output a flamegraph from a benchmark.
Unfortunately, this cannot happen during the same run as a standard
timing benchmark as it needs to plug-in a different profiler. [0]
Fortunately, we can use the same benchmarking code for both
types of profilers. The cli flag to switch between the two is
`cargo bench ... -- -- profile-time <seconds>`[1]

The flamegraph profiler comes from `pprof-rs`. Note that there is
dependency version conflist between `pprof-rs` and `rustyline` on
the `nix` library. `rustyline` is only used in the readyset-repl,
it's impact is tiny on the code base, so the required version of
`rustyline` was dropped by one to make it's `nix` dependency
compatible with `pprof`.

[0] https://bheisler.github.io/criterion.rs/book/user_guide/profiling.html
[1] <same url>#--profile-time

Change-Id: I79adda0cb54c67687c1240c733e112ef233b2c59
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6658
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>