nom-sql: Optional flamegraphs from benchmarks
Add functionality to output a flamegraph from a benchmark.
Unfortunately, this cannot happen during the same run as a standard
timing benchmark as it needs to plug-in a different profiler. [0]
Fortunately, we can use the same benchmarking code for both
types of profilers. The cli flag to switch between the two is
`cargo bench ... -- -- profile-time <seconds>`[1]

The flamegraph profiler comes from `pprof-rs`.

[0] https://bheisler.github.io/criterion.rs/book/user_guide/profiling.html
[1] https://bheisler.github.io/criterion.rs/book/user_guide/profiling.html#--profile-time

Change-Id: I3f9f3fa1a83c7e8b3bd2eccf863ce8c32281f8c1