readyset-psql: Fix flakey types::enum test
The `types::enum` test has been flakey for a while now in CI, and
(anecdotally) it seems to fail in the same place. The most most
likely issue is eventual consistency, as before the failing check,
the test drops the table and enum, recreates them, and a new cached
statement.

This patch adds an `eventually!` macro around this failing check
to give the replication additional time to catch up.

Fixes: REA-3635
Change-Id: Iafd15adcfad7f839501521b8af2a10228892e949