adapter: Add `DROP ALL PROXIED QUERIES`
This commit adds a new `DROP ALL PROXIED QUERIES` SQL extension, which
removes all the queries from the query status cache whose statuses are not
"Successful". This is being added to support end-to-end testing of the list
of queries in `SHOW PROXIED QUERIES` by allowing tests to clear the
queries in that list without having to restart ReadySet.

Change-Id: I58451eb11bf35227e34aec51854d64775c3d79f3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6663
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>