replicators: Allow alphanumeric replication server ids
It may be useful to use alphabetic characters in a replication server id
to uniquely identify a postgres replication slot--this commit expands
replication_server_id to allow for that.

Currently, mysql still requires a numeric server id, so that branch
still enforces that.

Release-Note-Core: REPLICATION_SERVER_ID can now be configured to use
  alphanumeric characters for postgres.

Change-Id: I6a9137f91d25029333f0f827e43ec099db6bbbf0