psql-src: Correcly handle ReadyForQuery messages
In the postges wire protocol, the ReadyForQuery backend message [0]
contains a trailing byte that indicates a status: idle, in a
transaction, or in a failed transaction (and must be aborted).
It's basically a transaction status flag. In all cases, however,
we set the status byte, as returned form ReadySet, to the IDLE flag.
This is incorrect when transactions are occurring.

This patch corrects that behavior by properly setting the transaction
status flag if we're in a transaction, or if we're in a transaction
that failed.

[0] https://www.postgresql.org/docs/current/protocol-message-formats.html

Fixes: REA-3988
Change-Id: I5e5f47d29b2749491bdef7180f90b0f624d932c0