readyset-adapter: Add timezone info to readyset status
Previously, `SHOW READYSET STATUS` provided times in a
ms-since-unix-epoch format, which implicitly have a time zone of UTC. We
changed that to format it in a more human-readable format, but that
meant that we no longer had the context of the time zone implicit in
unix timestamps.

This commit adds the context back by explicitly appending "UTC" to these
datetimes.

Change-Id: I0f152b30c7b76880cb9098c6f9934402f34f2232