df-expr: Fix convert_tz proptest
In the `convert_tz` proptest, we were not validating that the random
`NaiveDateTime`s being generated represented a valid time in the
"source" timezone being used in the test (Atlantic/Cape_Verde). This
commit adds a new strategy function to generate `NaiveDateTime`s that
represent valid times in the given timezone.

Change-Id: I7fc7d17e6216f334116b5942302013328d55e6ae
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6715
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>