adapter: add QueryResult::Deallocate variant
Adds a new QueryResult enum variant for Deallocate. The
existing enum variants, `Noria` and `Upstream`, are hints as
to where the response data came from. Deallocate is a bit trickier:
we need to start freeing a prepared statement higher in the stack,
but we can't know that until we actually parse the SQL statement
as DEALLOCATE comes in as a "query". Thus, the intent here is to
properly parse the SQL statement as normal, and have the adapter return a
"QueryResult" that the callers can execute higher-level functionality
(before probably calling into something like
`adapter::remove_statement()`).

Fixes: REA-3983
Change-Id: I79f348f491e3f4556bfdc6ec1c3a58703ab15d78