adapter: Optimize use of prepared_query_cache
`adapter::Backend::parse_query()` is called on both the adhoc query
and prepares paths. `parse_query()` checks a local
`parsed_query_cache` to see if it's parsed the incoming SQL
statement. The key into that cache is the literal SQL string itself.
In the case of adhoc queries, WHERE values are directly interpolated
into the string, like `SELECT * from t where id = 12345`. Thus,
there's a low chance of actually hitting a match in the cache as
most adhoc SQL stataments will be unique.

The real provlem is that values end up polluting the cache, and waste
a lot of memory, per connection. The cache is an `LruCache` instance,
and we currently hardcode an upper bound size of 10,000 elements.

For a workload that is dominated by adhoc queries, we use an excess
of memory with the cache, and then just thrash entries as there's
little query reuse (by the exact SQL string).

Change-Id: Ice50eea903eebc622bb6e2da410cb6f6999e26df
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6740
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>