adapter: Reconnect for upstream reachability test
The status reporter maintains a LazyUpstream connection to the upstream
to report whether the upstream is reachable or not. If the upstream
connection is severed (say, if postgres restarts), this connection was
no longer connected and reported unreachability. Now it attempts to
reconnect in this case and only reports unreachability if the
re-connection fails.

Fixes: REA-4039
Change-Id: Ia5d233f0effa3658cec822d7f02203e336e37ca5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6765
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>