metrics: Remove domain handle request time
This commit removes the metrics we were emitting that tracked the time
spend handling domain requests. These metrics were subject to a
cardinality explosion across the labels we're using, and ultimately, we
decided that this metric is not worth having for the time being. If we
need to, we can track the request times for individual domain request
variants to try to avoid the cardinality explosion.

Change-Id: I612073a9811b4fa81a6512f2c3c128e11e0b6cba
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6835
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI