replicators: Limit maximum snapshot parallelism
For smaller instances especially, snapshotting can saturate the CPU and
cause issues for things like background tasks, proxying, and external
health checks.

This commit allows for configuring the maximum number of parallel
snapshotting tables to allow. Note that the previous behavior seems to
have been limited to a parallelism of 50 by tokio itself.

Change-Id: I120c1035e20af461fd8624dd08c7dd11908fc0e3