readyset-server: configure persistence threads based on cpu
If we are running on an instance that has limited cpu cores, we dont
want to give too many of them to rocksdb, especially since during
snapshotting and compaction they can use quite a lot of CPU and cause
other things like health checks or proxied queries to experience
degradation.

This commit changes the default of persistnce_threads to be based on the
number of CPUS rather than a straight '6'.

Change-Id: I4c43c77e245c40a5c0cde7f9faaffb72ae3b9aa4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6840
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>