metrics: Rework startup metrics
Previously, we were using a counter metric to track the startup
timestamps of the adapter and server. Using a counter in this way makes
it difficult to graph the number of times Readyset has started up as a
time series. Really, all we have is a sum of a bunch of timestamps over
time, so the only way to get the number of startups over a given time
period is to count the number of *samples* of the counter metric. Since
metrics are already time series data, there's no real need to emit the
actual timestamp itself as a counter.

This commit removes the previous counter metric and replaces it with two
new ones, `READYSET_ADAPTER_STARTUPS` and `READYSET_SERVER_STARTUPS`,
which just count the number of startups for the adapter and the server,
respectively.

The `SHOW READYSET STATUS ADAPTER` command was relying upon the old
counter metric to return the process start time, though I believe there
may have been a bug here, since the unix timestamps were being summed to
produce this timestamp. To replace this, this commit also adds a proper
timestamp to `Backend` that is used in the `SHOW READYSET STATUS
ADAPTER` command.

Release-Note-Core: Simplified the metrics Readyset emits to count the
  number of times it starts up
Change-Id: I95e9c1e46a51fd377f8deb011ee4f5381b667679
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6849
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>