metrics: Add cache_name label
This commit adds a new `cache_name` label to a number of domain-related
metrics. This will allow us to group by the cache name when plotting
these metrics, making it much easier to determine whether the domains in
a cache are experiencing congestion due to slow replays.

Release-Note-Core: Added a new `cache_name` label to various metrics to
  improve debuggability
Change-Id: I44a4fd6760704ac0b9175990df745e2cc63a7325