adapter: Plumb through DeallocateId on Deallocate
Extend the ability to deallocate prepared statements by interpreting
the incoming statment identifier as something we've (probably) issued
(as a `u32`), as a name the client gave it (a string), and the ability
to deallocate any and all prepared statements.

Currently, If `adapter::Backend::remove_statement()` is requested to
close a prepared statement, and the `statement_id` is not recognized,
we ignore it. To support `DEALLOCATE` for statements that were
prepared via a `PREPARE` SQL statement (as we do not currently support
that, although the upstream database does), pass through a new enum,
`adapter-types::DeallocateId`, to the individual `UpstreamDatabase`
impls, where it can execute a `DEALLOCATE statement_id` query directly
to the upstream.

Fixes: REA-3983
Change-Id: I57b9e0a9bb23e492b12529cb0a28fe441718c2b6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6832
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>