adapter: Add QueryResult::Parser variant
Adds a new QueryResult enum variant for Parser commands which should
be handled by an outer layer of the call stack. The existing enum
variants, `Noria` and `Upstream`, are hints as to where the response
data came from. Sometimes, though, we parse a statement and determine
it should be reassessed at the outer layer.

This patch is motivated by needing to support `DEALLOCATE` because,
even though we successfullyt parse the statement, we need to start
freeing the prepared statement higher in the stack. Thus, the intent
here is to properly parse the SQL statement as normal, and have the
adapter return a `QueryResult` variant that the callers can interpret
to execute lower-level functionality (before probably calling into
something like `adapter::remove_statement()`).

This patch links up the previous patches in this CL to fully support
`DEALLOCATE` for both postgres and mysql. The patch cleans up random
caches of types, maps, and so on, hence the changes to support
dropping all data in the `channel` and `codec` modules in `psql-srv`.

Fixes: REA-3983
Release-Note-Core: Readyset now supports the
  `DEALLOCATE` SQL statement for both postgres
  and mysql.
Change-Id: I836dd4d671a87cef996816f4b6e6273b646a06b2