adapter: Transaction cmds as prepared statements
Readyset handles transaction boundaries correctly when it receives
`BEGIN`/`COMMIT`/`ROLLBACK` as simple, adhoc statements. However,
some drivers will send those statements as prepared statements, and
the adapter needs to handle those correctly, as well.

The downside of not handling this correctly, amongst other things, is
we do not return a correct `ReadyForQuery` message to the client.
Most prominently, we set the status to `I` (idle), when it should be
`T` (in transaction) when executing a BEGIN. Drivers like pgjdbc
check the ReadyForQuery status, and errors out if the status is not
as expected.

Fixes: REA-4072
Change-Id: Ic9bcaa64d0bf3fd0eaa031f780b520bb62a46a4e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6909
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>