build: Test demo w/ HEAD and current release
This commit updates our test-demo buildkite pipeline to test the demo
script and dashboard against the Readyset image associated with the
current `BUILDKITE_COMMIT` *in addition* to the last official release.
This allows us to make backwards incompatible changes to core such that
if the quickstart Docker compose files change, CI doesn't fail.

Change-Id: If822c7b1d27e1a1d70963c0223b420f66a7421d6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6875
Reviewed-by: Ron Hough <ron@readyset.io>
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI