adapter: Return upstream's CommandComplete Tag
When we proxy prepared statements to the upstream, if the query is
not a SELECT (meaning, we are not returning rows), we naively return
a count of rows inserted (`INSERT 0 0`) in our CommandCompleted
message back to the client. This is blatantly broken for UPDATE,
DELETE, BEGIN, COMMIT, and antything that is not an insert.

Part of this is due to the fact that the `rust-postgres` driver,
on the reponse of prepared statements, only returns "rows" or
a count of rows modified. This is reasonably sufficient for a driver,
but we need more of the upstream's wire protocol information, in order
to respond correct to our callers.

Once we get the CommandComplete Tag from the upstream, via a patch
to the `rust-postgres` driver (which must be committed before this
patch) [0], this patch does all the requisite plumbing to get that
Tag back to the caller.

[0] https://github.com/readysettech/rust-postgres/pull/13

Fixes: REA-4073
Change-Id: Ie8b3f5c94d9218c20aae8933649c6480b8a44d0c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6910
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>