demo: Update compose file name for Postgres
The compose.postgres.yml file has the name `readyset-postgres`, whereas
the BYO compose file has the name `readyset`. This causes discrepancies
when running certain commands, since the Readyset container name differs
based on the name of the compose file. This commit fixes the issue by
updating the name of the `compose.postgres.yml` file to `readyset` after
it's downloaded.

Change-Id: I6c955e16574c8672932586c165496a09d64a9222