metrics: Add on_input metrics
This commit adds metrics to track the duration and count of invocations
to `Ingredient::on_input`, labeled by the node operator type. This will
give us insight into what nodes are the most expensive nodes in the
graph and will help us spot anomalies (e.g. if a project node is taking
a very long time, something is probably wrong).

Change-Id: I85ace8d0d667a0c7d1e57100022bc053e8ec7cb0