replicators: Skip snapshotting tables in readyset schema
Our DDL triggers create a table in the readyset schema for versions of
postgres before 14, and this table was being snapshotted, which doesn't
make sense as this is an internal catalog table that is used to allow
for ddl replication to readyset.

This commit ignores tables in the `readyset` schema while determining
the set of tables to snapshot.

Fixes: REA-4077
Change-Id: I1dd03938d697d3b8327e7ce73fdc0e22fd5aadbd