psql-srv: Support pgjdbc Timestamp
During `BIND` for prepared statements, pgjdbc sends
`javax.sql.Timestamp`s with timezone data (UTC offset), even if the
target data type is `timestamp` (not `timestamptz`). According to
the PG docs [0], the timezone is optional, but Readyset errors if
there is timezone information in the string that is sent.
Interestingly, pgjdbc will send along only the hour offset,
truncating the minutes data, if there's no minutes offset for the
timezone (as is the case for most timezones). Note that this
"appending timezone data" applies to the text format transfer as
pgjdbc only transfers timestamps in the text format.

This patch extends the existing timestamp and timestamptz parsing by
checking for any trailing text in the input timestamp string, and
attempts to parse it into a timezone offset.

Also included is a benchmark that compares parsing the timezone
offset versus reparsing the entire input string with a "pgjdbc-
friendly" format string. Parsing just for the offset is 40% faster
than reparsing the entire string:

timestamp_text_parse/offsets
  time:   [281.25 ns 282.11 ns 283.06 ns]
timestamp_text_reparse/reparse
  time:   [485.31 ns 486.00 ns 486.81 ns]

[0] https://www.postgresql.org/docs/current/datatype-datetime.html

Fixes: REA-2662
Release-Note-Core: Support timestamps in prepared statements from
  the Java PG driver, pgjdbc.
Change-Id: Ie861d63d1c0ec8d13778857c18204c86585d834d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6932
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>