adapter+server: Autoparameterize range queries
This commit:
1. Adds support to the adapter autoparameterization pass to
   autoparameterize range queries (queries that include inequality
   comparisons); and
2. Updates the view key construction logic to support multiple
   inequality comparisons that can't be transformed into a
   `ViewPlaceholder::Between`

In the case of queries with mixed comparisons (those with equality
comparisons and inequality comparisons), autoparameterization will
succeed, but the `detect_unsupported_placeholders` pass will fail during
the server's AST rewrites.

I considered augmenting `ViewPlaceholder::Between` to include
inclusivity bounds (i.e. an enum to specify whether a bound was
inclusive or exclusive), but ultimately, I decided it made sense to keep
it as close to Postgres's `BETWEEN` semantics as possible, which always
assumes inclusive bounds.

There is further work here to deduplicate range queries that all rely on
the same index (e.g. `SELECT * FROM t WHERE x > $1 AND x < $10` and
`SELECT * FROM t WHERE x >= $1 AND x <= $10` can both use the same
reader node, since the only difference between them is the inclusivity
of the bounds). I started on this work but it seemed a bit more
challenging and figured it made sense to tackle it in a separate commit
anyway.

Fixes: REA-809
Release-Note-Core: Added support for the autoparameterization of range
  queries
Change-Id: Ifbe291fed2cd5562ff7d1d500f38d2473c387539