metrics: Add cache name to server view hit/miss metrics
This commit adds the cache name as a label to the counters we use to
track whether a query against a cache was a hit or a miss. Though we are
looking to keep label cardinality low, these metrics currently have no
other labels, and since they are counters, they don't have a "quantile"
label.

Release-Note-Core: Improved the metrics we use to track cache hits and
  misses
Change-Id: Id63866b793ef573baaa09bed9d464492b6540deb